Rename file crypto-lib to cryptoLib for naming consistency
